backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic-settings
python-dotenv
passlib[bcrypt]
PyJWT
